lt4rec/requirements.txt
tensorflow==1.12
tensorboard==1.12
pyyaml
absl-py
numpy
sklearn
